+++ src/agents/rl/true_online_sarsa/TrueOnlineSarsaLearner.hpp
@@ -17,17 +17,17 @@
 
 class TrueOnlineSarsaLearner : public RLLearner{
 	private:
-		double alpha, delta, lambda, traceThreshold;
+		float alpha, delta, lambda, traceThreshold;
 		int numFeatures, currentAction, nextAction;
 
 		std::string nameWeightsFile;
 
 		vector<int> F;					//Set of features active
 		vector<int> Fnext;              //Set of features active in next state
-		vector<double> Q;               //Q(a) entries
-		vector<double> Qnext;           //Q(a) entries for next action
-		vector<vector<double> > e;      //Eligibility trace
-		vector<vector<double> > w;      //Theta, weights vector
+		vector<float> Q;               //Q(a) entries
+		vector<float> Qnext;           //Q(a) entries for next action
+		vector<vector<float> > e;      //Eligibility trace
+		vector<vector<float> > w;      //Theta, weights vector
 		vector<vector<int> >nonZeroElig;//To optimize the implementation   
 
 		/**
@@ -43,10 +43,10 @@ class TrueOnlineSarsaLearner : public RLLearner{
 		/**
  		* In Sarsa the Q-values (one per action) are updated as the sum of weights for that given action.
  		* To avoid writing it more than once on the code, its update was extracted to a separate function.
- 		* It updates the vector<double> Q assuming that vector<int> F is filled, as it sums just the weights
+ 		* It updates the vector<float> Q assuming that vector<int> F is filled, as it sums just the weights
  		* that are active in F.
  		*/
-		void updateQValues(vector<int> &Features, vector<double> &QValues);
+		void updateQValues(vector<int> &Features, vector<float> &QValues);
 		/**
  		* When using Replacing traces, all values not related to the current action are set to 0, while the
  		* values for the current action that their features are active are set to 1. The traces decay following
@@ -54,9 +54,9 @@ class TrueOnlineSarsaLearner : public RLLearner{
  		*/
 		void decayTrace();
 
-		void updateTrace(int action, double alpha);
+		void updateTrace(int action, float alpha);
 
-		void updateWeights(int action, double alpha, double delta_q);
+		void updateWeights(int action, float alpha, float delta_q);
 		/**
  		* Prints the weights in a file. Each line will contain a weight.
  		*/
